Add CMake build files
The implementation does not include examples or tests, just the bare
library.

If the library is installed, it is expected that the consuming project
will be able to find_package(Gumbo CONFIG), target_link against
Gumbo::gumbo and include headers as <gumbo/gumbo.h>.

If the library is incorporated into the build directly (e.g. via
FetchContent), then it may be included with:

include(FetchContent)
FetchContent_Declare(
  gumbo-parser
  GIT_REPOSITORY https://github.com/hgkjshegfskef/gumbo-parser.git
  GIT_TAG cmake)
FetchContent_MakeAvailable(gumbo-parser)
target_link_libraries(<my_target> PRIVATE gumbo)

and the headers are available as <gumbo.h>.